Unify override transfer and let ButlerURI decide
Datastore no longer needs to know what "auto" transfer means
and it is best to leave it to ButlerURI unless we have the
special case of the files all being within datastore.
